Updated int() to float()

+++ main.py
@@ -48,29 +48,29 @@ gender_menu.place(relx=0.3, rely=0.4)
 
 def bmi_calc():
     try:
-        int(weight_entry.get())
-        int(height_entry.get())
-        int(age_entry.get())
+        float(weight_entry.get())
+        float(height_entry.get())
+        float(age_entry.get())
         if variable.get() == "Select...":
             raise ValueError
         elif variable.get() == "Male":
-            result = ((0.5 * int(weight_entry.get())) / ((int(height_entry.get()) / 100) ** 2)) + 11.5
+            result = ((0.5 * float(weight_entry.get())) / ((float(height_entry.get()) / 100) ** 2)) + 11.5
             result = round(result, 1)
             ideal_field.config(state='normal')
             ideal_field.insert(0, result)
             ideal_field.config(state='readonly')
-            result_bmi = int(weight_entry.get()) / ((int(height_entry.get()) / 100) ** 2)
+            result_bmi = float(weight_entry.get()) / ((float(height_entry.get()) / 100) ** 2)
             bmi_field.config(state='normal')
             bmi_field.insert(0, round(result_bmi, 1))
             bmi_field.config(state='readonly')
         elif variable.get() == "Female":
-            result = ((0.5 * int(weight_entry.get())) / ((int(height_entry.get()) / 100) ** 2)) + (
-                        0.03 * int(age_entry.get())) + 11
+            result = ((0.5 * float(weight_entry.get())) / ((float(height_entry.get()) / 100) ** 2)) + (
+                        0.03 * float(age_entry.get())) + 11
             result = round(result, 1)
             ideal_field.config(state='normal')
             ideal_field.insert(0, result)
             ideal_field.config(state='readonly')
-            result_bmi = int(weight_entry.get()) / ((int(height_entry.get()) / 100) ** 2)
+            result_bmi = float(weight_entry.get()) / ((float(height_entry.get()) / 100) ** 2)
             bmi_field.config(state='normal')
             bmi_field.insert(0, round(result_bmi, 1))
             bmi_field.config(state='readonly')
